+++ templates/comments.html
@@ -12,8 +12,8 @@
     {% endif %}
     {% for comment in post.comments|reverse %}
         <li><a href="{{ routes["user"].format(comment.author.username) }}">{{ comment.author.username }}</a>: {{ comment.content }}
-            {% if (user) and (user.role.value != "member") %}
-                            <a href="{{ routes["admin_comment_delete"].format(comment.comment_id) }}">
+            {% if (user) and ((user.role.value != "member") or (user.user_id == comment.author.user_id)) %}
+                            <a href="{{ routes["comment_delete"].format(comment.comment_id) }}">
                                 <img src="/static/images/x-solid-full.svg" width="10" height="10" alt="delete comment">
                             </a>
             {% endif %}
@@ -21,8 +21,8 @@
                 {% for reply in comment.replies %}
                     <li>
                         <a href="{{ routes["user"].format(reply.author.username) }}">{{ reply.author.username }}</a>: {{ reply.content }}
-                        {% if (user) and (user.role.value != "member") %}
-                            <a href="{{ routes["admin_comment_delete"].format(reply.comment_id) }}">
+                        {% if (user) and ((user.role.value != "member") or (user.user_id == reply.author.user_id)) %}
+                            <a href="{{ routes["comment_delete"].format(reply.comment_id) }}">
                                 <img src="/static/images/x-solid-full.svg" width="10" height="10" alt="delete comment">
                             </a>
                         {% endif %}
